# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API built with Express."
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Get API information",
        "responses": {
          "200": {
            "description": "Returns API information",
            "content": {
              "application/json": {
                "example": {
                  "name": "Task API",
                  "version": "1.0",
                  "endpoints": [
                    "/tasks"
                  ]
                }
              }
            }
          }
        }
      }
    },

    "/health": {
      "get": {
        "summary": "Check API health",
        "responses": {
          "200": {
            "description": "API is running",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok"
                }
              }
            }
          }
        }
      }
    },

    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "Returns all tasks",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "title": "Learn Express",
                    "done": false
                  },
                  {
                    "id": 2,
                    "title": "Learn Git",
                    "done": true
                  }
                ]
              }
            }
          }
        }
      },

      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Buy milk"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "example": {
                  "id": 4,
                  "title": "Buy milk",
                  "done": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },

    "/tasks/{id}": {
      "get": {
        "summary": "Get a task by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a single task"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },

      "put": {
        "summary": "Update a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "Master Git",
                "done": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Task not found"
          }
        }
      },

      "delete": {
        "summary": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "404": {
            "description": "Task not found"
          }
        }
      }
    }
  }
}